Do not validate platform if skipping Cordova build
I am attempting to set up hot code pushing and have the build and deployment process as part of CI.  As a result, I need to just build the Ember app without building the Cordova app and without having all of the Cordova platform / configuration / environment available.  The `--skip-cordova-build` option is great, but currently it still validates the platform and fails if it is not set up.  This platform validation seems unnecessary in this use case, and it is preventing the plain old Ember build from occurring.  The proposed change simply bypasses the platform validation when skipping the Cordova build.